webscrape/requirements.txt
beautifulsoup4==4.9.3
html-table-parser-python3==0.1.5
requests==2.25.1
```